requirements.txt
aniso8601==8.1.1
apispec==4.3.0
apispec-webframeworks==0.5.2
attrs==20.3.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
Flask==1.1.2

flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mysql-connector-python==8.0.25
protobuf==3.17.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
pyzk==0.9
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.5
Werkzeug==1.0.1
